+++ src/components/MdDatepicker/MdDatepicker.vue
@@ -1,12 +1,18 @@
 <template>
   <md-field :class="['md-datepicker', { 'md-native': !this.mdOverrideNative }]" md-clearable>
     <md-date-icon class="md-date-icon" @click.native="toggleDialog" />
-    <md-input :type="type" ref="input" v-model="modelDate" @focus.native="onFocus" pattern="[0-9]{4}-[0-9]{2}-[0-9]{2}" />
+    <md-input :type="type" ref="input" :value="modelDate" @input="onInput" @focus.native="onFocus" pattern="[0-9]{4}-[0-9]{2}-[0-9]{2}" />
 
     <slot />
 
     <keep-alive>
-      <md-datepicker-dialog :md-date.sync="selectedDate" :md-disabled-dates="mdDisabledDates" v-if="showDialog" @md-closed="toggleDialog" />
+      <md-datepicker-dialog
+        v-if="showDialog"
+        :md-date.sync="selectedDate"
+        :md-disabled-dates="mdDisabledDates"
+        :mdImmediately="mdImmediately"
+        @md-closed="toggleDialog"
+      />
     </keep-alive>
 
     <md-overlay class="md-datepicker-overlay" md-fixed :md-active="showDialog" @click="toggleDialog" />
@@ -22,6 +28,7 @@
   import MdOverlay from 'components/MdOverlay/MdOverlay'
   import MdDatepickerDialog from './MdDatepickerDialog'
   import MdDateIcon from 'core/icons/MdDateIcon'
+  import MdDebounce from 'core/utils/MdDebounce'
   import MdField from 'components/MdField/MdField'
   import MdInput from 'components/MdField/MdInput/MdInput'
 
@@ -44,6 +51,14 @@
       mdOverrideNative: {
         type: Boolean,
         default: true
+      },
+      mdImmediately: {
+        type: Boolean,
+        default: false
+      },
+      MdDebounce: {
+        type: Number,
+        default: 1000
       }
     },
     data: () => ({
@@ -77,10 +92,18 @@
           if (isValid(parsedDate)) {
             this.selectedDate = parsedDate
           }
+        } else {
+          this.selectedDate = null
         }
       }
     },
     methods: {
+      onInput(value) {
+        const parsedDate = parse(value)
+        if (isValid(parsedDate)) {
+          this.selectedDate = parsedDate
+        }
+      },
       toggleDialog () {
         if (!isFirefox || this.mdOverrideNative) {
           this.showDialog = !this.showDialog
@@ -114,6 +137,7 @@
       }
     },
     created () {
+      this.onInput = MdDebounce(this.onInput, this.MdDebounce)
       this.modelDate = this.dateToHTMLString(this.value)
       this.selectedDate = this.value
     }