+++ src/components/MdDrawer/MdDrawer.vue
@@ -20,7 +20,6 @@
       MdOverlay
     },
     props: {
-      mdLeft: Boolean,
       mdRight: Boolean,
       mdPermanent: {
         type: String,
@@ -57,7 +56,7 @@
     computed: {
       drawerClasses () {
         let classes = {
-          'md-left': this.mdLeft,
+          'md-left': !this.mdRight,
           'md-right': this.mdRight,
           'md-temporary': this.isTemporary,
           'md-persistent': this.mdPersistent,
@@ -176,8 +175,16 @@
     }
 
     &.md-temporary {
-      + .md-app-container .md-content {
-        border-left: none;
+      &.md-left {
+        + .md-app-container .md-content {
+          border-left: none;
+        }
+      }
+
+      &.md-right-previous {
+        + .md-app-container .md-content {
+          border-right: none;
+        }
       }
 
       &.md-active {
@@ -227,22 +234,45 @@
 
     &.md-persistent {
       &:not(.md-active) {
-        + .md-app-container .md-content {
-          border-left: none;
+        &.md-left {
+          + .md-app-container .md-content {
+            border-left: none;
+          }
+        }
+
+        &.md-right-previous {
+          + .md-app-container .md-content {
+            border-right: none;
+          }
         }
       }
     }
 
     &.md-persistent-mini {
-      border-right: 1px solid;
       transform: translate3D(0, 64px, 0);
       transition: .3s $md-transition-stand-timing;
       transition-property: transform, width;
       will-change: transform, box-shadow;
 
+      &.md-left {
+        border-right: 1px solid;
+      }
+
+      &.md-right {
+        border-left: 1px solid;
+      }
+
       &.md-active {
-        + .md-app-container .md-content {
-          border-left: none;
+        &.md-left {
+          + .md-app-container .md-content {
+            border-left: none;
+          }
+        }
+
+        &.md-right-previous {
+          + .md-app-container .md-content {
+            border-right: none;
+          }
         }
       }