+++ src/components/MdTable/MdTable.vue
@@ -27,8 +27,8 @@
         <tbody v-else-if="value.length">
           <md-table-row-ghost
             v-for="(item, index) in value"
-            :key="getRowId(item[mdModelId])"
-            :md-id="getRowId(item[mdModelId])"
+            :key="getRowId(item, mdModelId)"
+            :md-id="getRowId(item, mdModelId)"
             :md-index="index"
             :md-item="item">
             <slot name="md-table-row" :item="item" />
@@ -47,7 +47,7 @@
       <slot name="md-table-pagination" />
     </md-content>
 
-    <slot v-if="value" />
+    <slot v-if="!hasValue && $scopedSlots['md-table-row']" />
   </md-tag-switcher>
 </template>
 
@@ -153,7 +153,8 @@
           getModel: this.getModel,
           getModelItem: this.getModelItem,
           selectingMode: null
-        }
+        },
+        itemsUuidMap: new WeakMap()
       }
     },
     computed: {
@@ -234,16 +235,17 @@
       },
       'MdTable.selectedItems' (val, old) {
         let changed = (() => {
-          let isValEmpty = !val || val.length === 0
-          let isOldEmpty = !old || old.length === 0
+          let isValEmpty = this.isEmpty(val)
+          let isOldEmpty = this.isEmpty(old)
+          let hasValues = isValEmpty && isOldEmpty
 
-          if (isValEmpty && isOldEmpty) {
+          if (hasValues) {
             return false
-          } else if (!isValEmpty && !isOldEmpty) {
+          } else if (!hasValues) {
             return (val.length !== old.length) ? true : !val.every((item, index) => item == old[index])
-          } else {
-            return true
           }
+
+          return true
         })()
 
         if (changed) {
@@ -260,15 +262,27 @@
       }
     },
     methods: {
+      isEmpty (value) {
+        return !value || value.length === 0
+      },
       emitEvent (eventName, value) {
         this.$emit(eventName, value)
       },
-      getRowId (id) {
+      getRowId (item, propertyName) {
+        let id = item[propertyName]
+
         if (id) {
           return id
         }
 
-        return 'md-row-' + MdUuid()
+        id = this.itemsUuidMap.get(item)
+
+        if (!id) {
+          id = 'md-row-' + MdUuid()
+          this.itemsUuidMap.set(item, id)
+        }
+
+        return id
       },
       setScroll ($event) {
         raf(() => {