+++ src/components/MdTabs/MdTabs.vue
@@ -208,7 +208,7 @@
 
               this.indicatorStyles = {
                 left: `${buttonLeft}px`,
-                width: `${buttonWidth}px`
+                right: `calc(100% - ${buttonWidth + buttonLeft}px)`
               }
             }
           })
@@ -310,8 +310,11 @@
       will-change: height;
     }
 
-    &.md-transparent .md-tabs-navigation {
-      background-color: transparent !important;
+    &.md-transparent {
+      .md-tabs-navigation,
+      .md-tabs-content {
+        background-color: transparent !important;
+      }
     }
 
     &.md-dynamic-height .md-tabs-content {